refactor(match): some methods added and fixed
