+++ src/bitrule/practice/match/impl/TeamMatchImpl.php
@@ -8,6 +8,7 @@ use bitrule\practice\match\AbstractMatch;
 use bitrule\practice\match\Team;
 use bitrule\practice\profile\DuelProfile;
 use pocketmine\player\Player;
+use pocketmine\world\Position;
 use function array_merge;
 use function ceil;
 use function count;
@@ -20,14 +21,44 @@ final class TeamMatchImpl extends AbstractMatch {
     /**
      * @param Player $player
      */
-    public function teleportSpawn(Player $player): void {
-        // TODO: Implement teleportSpawn() method.
+    public function joinSpectator(Player $player): void {
+        if (($team = $this->teams[2] ?? null) === null) {
+            $this->teams[2] = $team = new Team(2, []);
+        }
+
+        $team->addPlayer($player->getXuid());
+
+        $this->postJoinSpectator($player);
+    }
+
+    /**
+     * Get the spawn id of the player
+     * If is single match the spawn id is the index of the player in the players array.
+     * If is team match the spawn id is the team id of the player.
+     *
+     * @param string $xuid
+     *
+     * @return int
+     */
+    public function getSpawnId(string $xuid): int {
+        foreach ($this->teams as $team) {
+            if (!$team->isMember($xuid)) continue;
+
+            return $team->getId();
+        }
+
+        return -1;
     }
 
     /**
      * @param Player $player
+     * @param bool   $canEnd
      */
-    public function removePlayer(Player $player): void {
+    public function removePlayer(Player $player, bool $canEnd): void {
+        if ($canEnd && count($this->getAlive()) <= 1) {
+            $this->end();
+        }
+
         foreach ($this->teams as $team) {
             if (!$team->removePlayer($player->getXuid())) continue;
 
@@ -69,5 +100,21 @@ final class TeamMatchImpl extends AbstractMatch {
         }
 
         $this->teams = $teams;
+    }
+
+    /**
+     * This method is called when the match stage change to Ending.
+     * Usually is used to send the match results to the players.
+     */
+    public function end(): void {
+        throw new \RuntimeException('Not implemented');
+    }
+    /**
+     * @param string $xuid
+     *
+     * @return string|null
+     */
+    public function getOpponentName(string $xuid): ?string {
+        throw new \RuntimeException('Not implemented');
     }
 }
